backend/requirements.txt
acme==1.21.0
asgiref==3.6.0
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.1.2
Automat==22.10.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
backports.zoneinfo==0.2.1
black==22.6.0
boto3==1.34.58
botocore==1.34.58
certbot==1.21.0
certbot-nginx==1.21.0
chardet==4.0.0
# cloud-init==23.4
colorama==0.4.4
# command-not-found==0.3
ConfigArgParse==1.5.3
configobj==5.0.6
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
coverage==6.5.0
coveralls==3.3.1
cryptography==42.0.5
# dbus-python==1.2.18
distlib==0.3.8
distro==1.7.0
# distro-info==1.1+ubuntu0.2
Django==3.2
django-cors-headers==4.3.1
django-rest-framework==0.1.0
daphne==4.1.0
django-ipware==6.0.4
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docopt==0.6.2
et-xmlfile==1.1.0
filelock==3.13.1
flake8==4.0.1
gevent==21.8.0
greenlet==1.1.2
gunicorn==21.2.0
# gyp==0.1
httplib2==0.20.2
hyperlink==21.0.0
idna==3.6
import-export==0.3.1
importlib-metadata==4.6.4
incremental==22.10.0
jmespath==1.0.1
jeepney==0.7.1
Jinja2==3.0.3
josepy==1.10.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.0.1
more-itertools==8.10.0
mccabe==0.6.1
msgpack==1.0.8
mypy-extensions==1.0.0
ndg-httpsclient==0.5.1
netifaces==0.11.0
numpy==1.24.4
openpyxl==3.1.2
oauthlib==3.2.0
parsedatetime==2.6
pexpect==4.8.0
psycopg2==2.9.2
ptyprocess==0.7.0
# PyGObject==3.42.1
PyHamcrest==2.0.2
# PyICU==2.8.1
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
pyserial==3.5
# python-apt==2.4.0+ubuntu3
# python-debian==0.1.43+ubuntu1.1
python-magic==0.4.24
# PyYAML==5.4.1
packaging==23.2
pandas==2.0.3
pathspec==0.12.1
platformdirs==4.2.0
psycopg2-binary==2.9.9
pusher==3.3.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==24.0.0
python-dateutil==2.9.0.post0
python-ipware==2.0.2
pytz==2022.7.1
redis==5.0.2
requests==2.31.0
# sos==4.5.6
sqlparse==0.4.4
ssh-import-id==5.11
# systemd-python==234
s3transfer==0.10.0
service-identity==24.1.0
six==1.16.0
tablib==3.5.0
tomli==2.0.1
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
zope.interface==6.2
# ubuntu-advantage-tools==8001
# ufw==0.36.1
# unattended-upgrades==0.1
virtualenv==20.25.1
wadllib==1.3.6
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0